feat: Emphasis
Due to the nesting of emphasis lexemes, we need to be careful with the
amount of semantic interpretion performed by the lexer.

The col property keeps track of the column position of the line,
avoiding issues where the "." scene heading handling is incorrectly
tokenized due to a broken period in a text block. For example,
"_emphasized text_." will float a dangling period back into the
tokenizer, catching the scene heading branch.

Newlines can be enforced by looking at l.col == 0. This will probably
come in handy when dealing with nexting edge-cases. e.g. a sentence
including "EXT. FOO BAR" may be flagged as a scene heading, even if it
isn't at the beginning of the line. It's likely that we'll want to add
l.col checks to each of the newline-enforced lexemes.